data단에서 래퍼로 감싸 도메인으로 내리기
